@@ -17,19 +17,18 @@
     }
   ],
   "require": {
-    "php": "^7.1.3",
+    "php": "^7.4",
     "ext-ctype": "*",
     "ext-iconv": "*",
     "ext-json": "*",
     "a2lix/translation-form-bundle": "^3.0",
-    "barthy/cached-prezent-translation": "^1.0",
-    "cocur/slugify": "^3.2",
+    "barthy-koeln/cached-prezent-translation": "^1.0",
+    "cocur/slugify": "^3.2 || ^4.0",
     "liip/imagine-bundle": "^2.1",
     "prezent/doctrine-translatable-bundle": "^1.0",
     "stof/doctrine-extensions-bundle": "^1.3",
     "symfony/form": "^4.4|^5.0",
-    "symfony/orm-pack": "^1.0",
-    "symfony/validator": "^4.4|^5.0",
+    "symfony/validator": "^4.4 || ^5.0",
     "twig/twig": "^2.11 || ^3.0",
     "vich/uploader-bundle": "^1.8"
   },
@@ -39,40 +38,54 @@
     },
     "sort-packages": true
   },
-  "repositories": [
-    {
-      "type": "vcs",
-      "url": "https://github.com/barthy-koeln/cached-prezent-translation.git"
-    }
-  ],
   "autoload": {
     "psr-4": {
-      "Barthy\\ImageUploadBundle\\": "src/"
+      "BarthyKoeln\\ImageUploadBundle\\": "src/"
     }
   },
   "autoload-dev": {
     "psr-4": {
-      "Barthy\\ImageUploadBundle\\Tests\\": "tests/"
+      "Tests\\": "tests/"
     }
   },
   "conflict": {
     "symfony/symfony": "*"
   },
+  "extra": {
+    "hooks": {
+      "pre-commit": [
+        "bin/pre-commit"
+      ]
+    }
+  },
   "minimum-stability": "dev",
   "prefer-stable": true,
   "require-dev": {
+    "brainmaestro/composer-git-hooks": "^2.8",
+    "codeception/assert-throws": "^1.1",
     "dama/doctrine-test-bundle": "^6.2",
-    "phpunit/phpunit": "^7.5",
+    "doctrine/doctrine-bundle": "^2.2",
+    "doctrine/orm": "^2.7",
+    "friendsofphp/php-cs-fixer": "^2.16",
+    "phpmd/phpmd": "^2.7",
+    "phpunit/phpunit": "^9.4",
     "sensio/framework-extra-bundle": "^5.1",
-    "symfony/asset": "^4.1",
-    "symfony/console": "^4.1",
-    "symfony/expression-language": "^4.1",
+    "squizlabs/php_codesniffer": "^3.5",
+    "symfony/asset": "^5.1",
+    "symfony/console": "^5.1",
+    "symfony/expression-language": "^5.1",
     "symfony/flex": "^1.0",
     "symfony/monolog-bundle": "^3.1",
-    "symfony/process": "^4.1",
-    "symfony/translation": "^4.1",
-    "symfony/twig-bundle": "^4.1",
-    "symfony/web-link": "^4.1",
-    "symfony/yaml": "^4.1"
+    "symfony/process": "^5.1",
+    "symfony/translation": "^5.1",
+    "symfony/twig-bundle": "^5.1",
+    "symfony/web-link": "^5.1",
+    "symfony/yaml": "^5.1"
+  },
+  "scripts": {
+    "auto-scripts": {
+      "cache:clear": "symfony-cmd",
+      "assets:install %PUBLIC_DIR%": "symfony-cmd"
+    }
   }
 }
