@@ -1,20 +1,17 @@
 <?php
 
-namespace Barthy\ImageUploadBundle\DependencyInjection;
+namespace BarthyKoeln\ImageUploadBundle\DependencyInjection;
 
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\Extension;
 use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
 
-class BarthyImageUploadExtension extends Extension
+class BarthyKoelnImageUploadExtension extends Extension
 {
-
     /**
      * Loads a specific configuration.
      *
-     * @param array $configs
-     * @param ContainerBuilder $container
      * @throws \Exception
      */
     public function load(array $configs, ContainerBuilder $container)
@@ -22,13 +19,13 @@ class BarthyImageUploadExtension extends Extension
         // Process internal config files
         $yamlLoader = new YamlFileLoader(
             $container,
-            new FileLocator(__DIR__."/../Resources/config")
+            new FileLocator(__DIR__.'/../Resources/config')
         );
 
-        $yamlLoader->load("services.yaml");
+        $yamlLoader->load('services.yaml');
 
         $configuration = new Configuration();
-        $config = $this->processConfiguration($configuration, $configs);
+        $config        = $this->processConfiguration($configuration, $configs);
 
         $container->getDefinition(ImageUploadConfig::class)->setArgument(0, $config);
     }
