@@ -1,20 +1,12 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: Barthy
- * Date: 11.07.18
- * Time: 17:18
- */
 
-namespace Barthy\ImageUploadBundle\DependencyInjection;
+namespace BarthyKoeln\ImageUploadBundle\DependencyInjection;
 
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
 
-
 class Configuration implements ConfigurationInterface
 {
-
     /**
      * Generates the configuration tree builder.
      *
@@ -22,22 +14,22 @@ class Configuration implements ConfigurationInterface
      */
     public function getConfigTreeBuilder()
     {
-        $treeBuilder = new TreeBuilder('barthy_image_upload');
+        $treeBuilder = new TreeBuilder('barthy_koeln_image_upload');
 
         $treeBuilder->getRootNode()
             ->children()
-                ->scalarNode('file_name_language')
-                    ->defaultValue('en')
-                ->end()
-                ->scalarNode('max_file_size')
-                    ->defaultValue('2M')
-                ->end()
-                ->scalarNode('image_class')
-                    ->isRequired()
-                ->end()
-                ->scalarNode('image_path_prefix')
-                    ->isRequired()
-                ->end()
+            ->scalarNode('file_name_language')
+            ->defaultValue('en')
+            ->end()
+            ->scalarNode('max_file_size')
+            ->defaultValue('2M')
+            ->end()
+            ->scalarNode('image_class')
+            ->isRequired()
+            ->end()
+            ->scalarNode('image_path_prefix')
+            ->isRequired()
+            ->end()
             ->end();
 
         return $treeBuilder;
