@@ -1,29 +1,22 @@
 <?php
 
+namespace Tests\Entity;
 
-namespace Barthy\ImageUploadBundle\Tests\Entity;
-
-
-use Barthy\ImageUploadBundle\Entity\ImageTranslationTrait;
+use BarthyKoeln\ImageUploadBundle\Entity\ImageTranslationTrait;
 use Doctrine\ORM\Mapping as ORM;
 use Prezent\Doctrine\Translatable\Annotation as Prezent;
 use Prezent\Doctrine\Translatable\Entity\AbstractTranslation;
 
 /**
- * Class SpecificImageTranslation
- *
  * @ORM\Entity()
- *
- * @package Barthy\ImageUploadBundle\Tests\Entity
  */
 class SpecificImageTranslation extends AbstractTranslation
 {
-
     use ImageTranslationTrait;
 
     /**
-     * @var \Barthy\ImageUploadBundle\Tests\Entity\SpecificImage
-     * @Prezent\Translatable(targetEntity="\Barthy\ImageUploadBundle\Tests\Entity\SpecificImage")
+     * @var \Tests\Entity\SpecificImage
+     * @Prezent\Translatable(targetEntity="Tests\Entity\SpecificImage")
      */
     protected $translatable;
 }
