@@ -1,26 +1,22 @@
 <?php
 
-namespace Barthy\ImageUploadBundle\Test\Validator;
+namespace Tests\Validator;
 
-use Barthy\ImageUploadBundle\DependencyInjection\ImageUploadConfig;
-use Barthy\ImageUploadBundle\Entity\ImageTranslationTrait;
-use Barthy\ImageUploadBundle\Tests\Entity\SpecificImage;
-use Barthy\ImageUploadBundle\Tests\Entity\SpecificImageTranslation;
-use Barthy\ImageUploadBundle\Validator\FileTitleConstraint;
-use Barthy\ImageUploadBundle\Validator\FileTitleConstraintValidator;
+use BarthyKoeln\ImageUploadBundle\DependencyInjection\ImageUploadConfig;
+use BarthyKoeln\ImageUploadBundle\Validator\FileTitleConstraint;
+use BarthyKoeln\ImageUploadBundle\Validator\FileTitleConstraintValidator;
 use ReflectionClass;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use Symfony\Component\Validator\Constraint;
+use Tests\Entity\FileSizeEntity;
+use Tests\Entity\SpecificImage;
+use Tests\Entity\SpecificImageTranslation;
 
 class FileTitleConstraintTest extends KernelTestCase
 {
+    private FileTitleConstraint $constraint;
 
-    /**
-     * @var FileTitleConstraint
-     */
-    private $constraint;
-
-    public function setUp()
+    public function setUp(): void
     {
         self::bootKernel();
 
@@ -30,34 +26,29 @@ class FileTitleConstraintTest extends KernelTestCase
     }
 
     /**
-     * @covers \Barthy\ImageUploadBundle\Validator\FileTitleConstraint::validatedBy
-     * @covers \Barthy\ImageUploadBundle\Validator\FileTitleConstraint::getTargets
-     * @covers \Barthy\ImageUploadBundle\Validator\FileTitleConstraintValidator::validate
      * @throws \Exception
      */
     public function testValidation()
     {
         $validator = self::$container->get('validator');
-        $entity = new SpecificImage();
+        $entity    = new SpecificImage();
 
         $violations = $validator->validate($entity, $this->constraint);
         $this->assertEquals(1, $violations->count());
 
-        /**
-         * @var  ImageTranslationTrait $translation
-         */
         $translation = new SpecificImageTranslation();
         $translation->setTitle('something');
         $translation->setLocale('de');
 
         $entity->addTranslation($translation);
         $violations = $validator->validate($entity, $this->constraint);
+        $this->assertEquals(0, $violations->count());
 
+        $violations = $validator->validate(new FileSizeEntity(), $this->constraint);
         $this->assertEquals(0, $violations->count());
     }
 
     /**
-     * @covers \Barthy\ImageUploadBundle\Validator\FileTitleConstraintValidator::__construct
      * @throws \ReflectionException
      */
     public function testInjection()
@@ -65,11 +56,11 @@ class FileTitleConstraintTest extends KernelTestCase
         /**
          * @var ImageUploadConfig $config
          */
-        $config = self::$container->get(ImageUploadConfig::class);
+        $config    = self::$container->get(ImageUploadConfig::class);
         $validator = new FileTitleConstraintValidator($config);
 
         $reflection = new ReflectionClass($validator);
-        $property = $reflection->getProperty('config');
+        $property   = $reflection->getProperty('config');
         $property->setAccessible(true);
 
         self::assertEquals($config, $property->getValue($validator));
